pyproject.toml
[tool.poetry]
name = "TasmotaToInflux"
version = "0.0.1"
description = ""
authors = [
    "David Bauer <github@debauer.net>",
]
license = "MIT"
readme = "README.md"

packages = [
    { include = "tasmotatoinflux", from = "src" }
]

[tool.poetry.scripts]
tams = "tasmotatoinflux.core:core"

[tool.poetry.dependencies]
python = "^3.10"
influxdb = "^5.3.1"
overrides = "^7.3.1"
flatten-json = "^0.1.14"
python-json-config = "^1.2.3"
paho-mqtt = "^1.6.1"
ruff = "^0.1.5"

[tool.poetry.dev-dependencies]
types-requests = "^2.27.30"
black = "^22.3.0"
mypy = "^0.961"
pylint = "^2.14.2"
pyupgrade = "^2.34.0"
isort = "^5.10.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry_core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
files = "src"

junit_xml = "xunit-reports/xunit-result-mypy.xml"

# output style configuration
show_column_numbers = true
show_error_codes = true
pretty = true

# additional warnings
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_no_return = true

no_implicit_optional = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
# disallow_any_explicit = true
disallow_any_generics = true
disallow_untyped_calls = true

enable_error_code = [
    "ignore-without-code",
]

# ignores that library has no typing information with it
[[tool.mypy.overrides]]
module = [
    "paho.*",
    "influxdb",
    "tasmotatoinflux.config",
]
ignore_missing_imports = true


[tool.ruff]
select = ["ALL", "PLR6301"]
preview = true
ignore = [
  "D1",      # docstrings
  "D406",    # weird section heading things
  "D203",    # one-blank-line-before-class
  "D212",    # multi-line-summary-first-line
  "ERA001",  # Commented out code found
  "PLC0414", # useless import alias
  "TCH",     # type checking blocks
  "TD003",   # todo issue links
  "ANN10",   # typing self and cls
  "FIX002",  # TODO anywhere in comments
  "CPY001",  # Copyright notices until configurable
  "A003",
  "E203",    # space before ':' has some inconsitencies with black
]

line-length = 120
target-version = "py38"

[tool.ruff.per-file-ignores]
"src/*" = ["PT"] # pytest rule set

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["risk_model.model.types.Sigma"]

[tool.ruff.isort]
force-single-line = true
lines-between-types = 1
lines-after-imports = 2
known-first-party = ["risk_model"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"
